Add html support for search engine
Added three new files that help in transitioning to a web like
interface. Also slightly modified the inverted index to go along

Still takes a lot of time for a query and may output memory error
because it loads the entire index in memory for every query.
Optimizations should be made