Improve query speed by a lot and clean some code
Further improvements have been made to reduce the time needed from the
time the user presses search to the time the results are showed in the
screen. Also a lot of test code has been cleaned. Now the user needs to
run create_catalog.py to create the database and the catalog in the
database and then to run the app.py file so they can execute queries in
a webpage.

Improvements can still be made by handling unknown terms in the query