Support top-k similarity between politicians
Through a 3 staged code in parliament_members_similarities.py we can
find the top-k most similar pair of politicians. This was done by
expressing politicians by a feature vector which consists of their last
name, all of the parties that they participated in, the regions that
they represented, the roles they had in each goverment, their gender
and lastly their most important keywords (which are extracted with
the k-core decomposition). Jaccard similarity was used for the
comparisons and every feature had a weight. For example the gender is
not that important as the keywords.