For every goverment find the normalized number of words for each party
Essentially see how much each party talks in the parliament regarding
the number of politicians they have. This was done by repeated sql
queries

The date field in the database cannot work in a range query so the id
field is used instead. This means that the ids are picked manually. An
improvement can be made for automatically finding the ids that separate
each goverment